feat : 물 등록, 체중등록 기능 수정
